pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentproof"
version = "0.1.0"
description = "AI agent observability, benchmarking, and attestation platform"
requires-python = ">=3.12"
dependencies = [
    "litellm>=1.61.0,<2.0",
    "asyncpg>=0.30.0",
    "sqlalchemy[asyncio]>=2.0",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "orjson>=3.10",
    "typer>=0.15",
    "rich>=13.0",
    "httpx>=0.28",
    "scipy>=1.14",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.25",
    "pytest-cov>=6.0",
    "testcontainers[postgres]>=4.0",
    "locust>=2.32",
    "ruff>=0.9",
    "mypy>=1.14",
    "pre-commit>=4.0",
]
ml = [
    "onnxruntime>=1.20",
    "transformers>=4.47",
    "datasets>=3.2",
]

[project.scripts]
agentproof = "agentproof.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/agentproof"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-ra -q"
markers = [
    "integration: tests requiring Docker (TimescaleDB testcontainer)",
    "framework: tests requiring live LiteLLM proxy and framework SDKs (langchain, crewai)",
]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "TCH"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]
